TILA-1078 save image changes with rest of the changes
